Fix : auctionId, lotId 사용 쿼리 에러 해결
