+++ gokkan/src/components/PaymentResult/OrderNumber.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import styled from 'styled-components';
-import Section from '../Payment/Section';
 
 interface IProps {
   orderNumber: string;
